Move generateAnkiNoteMediaFileName to anki-util
